Rename ProxyServer.scheme to .protocol.
Originally ProxyServer.scheme is named according to the Chrome Extension
API, but it turns out to be causing confusion with the scheme of
proxified URLs. The UI tries to address this by declaring this field as
"protocol". The code should also do so.
Proxies have "protocols" and URLs have "schemes". It makes perfect sense.