Add editors for FixedProfile. Redesign settings UI for FixedProfile.
The "Use same proxies for all protocols" checkbox is gone.
A new option is added to the Protocol combobox called "(use default)",
which applies the default proxy to the current scheme.
Friendly placeholders are used in host and port fields to indicate the
default value (or the value of the default proxy, if "use default" is
selected.)