Add confirm dialog for resetting rules.
And:
Fix a bug in Communicator causing replies to be misinterpreted as
actions when reply callbacks are not set.
Fix another bug in SwitchProfile. The name of the defaultProfileName
setter is now corrected.
Move restoring previous opened tab function to .dart instead of .js,
which fast enough to prevent tab-jumping on initialization.
Show a splash screen when options are not loaded.